Enable releasing of npm artifacts into artifactory from pipeline
Added java library README.md
