+++ imports-frontend-entities/.npmrc
@@ -1,0 +1,1 @@
+@ipaffs:registry=https://artifactory.azure.defra.cloud/artifactory/api/npm/npm-snapshots-local/
