+++ imports-frontend-entities/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@artifactory/imports-frontend-entities",
-  "version": "1.0.6",
+  "name": "@ipaffs/imports-frontend-entities",
+  "version": "1.0.21",
   "private": false,
   "description": "Baseline schema and entities to use across all frontends",
   "files": [
@@ -9,10 +9,13 @@
     "src/entities/*.js"
   ],
   "scripts": {
-    "prepublish": "cp ../notification-schema-core/resources/notification-schema.json ./etc/"
+    "prepare": "cp ../notification-schema-core/resources/notification-schema.json ./etc/",
+    "test": "nyc mocha --recursive",
+    "release": "release-it --ci --no-git",
+    "login": "npm-cli-login -u $ARTIFACTORY_USERNAME -p $ARTIFACTORY_PASSWORD -e imports-jenkins@defra.onmicrosoft.com -r https://artifactory.azure.defra.cloud/artifactory/api/npm/npm-releases-local -s @ipaffs"
   },
   "publishConfig": {
-    "registry": "https://artifactory.azure.defra.cloud/artifactory/api/npm/npm-virtual-snapshots/"
+    "registry": "https://artifactory.azure.defra.cloud/artifactory/api/npm/npm-snapshots-local/"
   },
   "author": "",
   "license": "ISC",
@@ -29,6 +32,8 @@
   "dependencies": {
     "ajv": "6.5.3",
     "lodash": "4.17.11",
-    "nyc": "14.1.1"
+    "nyc": "14.1.1",
+    "release-it": "12.3.0",
+    "npm-cli-login": "0.1.1"
   }
 }
